0.78 Progress 2
BUG - Fix camera at combat start showing nothing. BUG - Fix map scroll not starting at bottom. BUG - Fixed combat skills blocking UI in map. BUG - details bottom UI shows during item room rewards, it shouldnt. BUG - After a combat, shop view is moved, it shouldnt. BUG - Ensure speed on pc (testing) for atk minigame is same for mobile

+++ Assets/Scripts/HeroWeapon.cs
@@ -5,7 +5,7 @@ using UnityEngine.UI;
 using DG.Tweening;
 public class HeroWeapon : MonoBehaviour
 {
-    public int lineSpeed;
+    public float lineSpeed;
     public string ownedUnitName;
     public Transform hitLine;
     public Transform topBarBorder;
